Perf filter.
